% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bootmeans2adjacency}
\alias{bootmeans2adjacency}
\title{create matrix of probabilities i>j}
\usage{
bootmeans2adjacency(meanvotes.boot.list, names = NULL)
}
\description{
create matrix of probabilities i>j
}
