% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data2longlist}
\alias{data2longlist}
\title{convert cleaned data into long form}
\usage{
data2longlist(data, vote.colnames, title.colname, type.colname)
}
\description{
convert cleaned data into long form
}
