% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filmrank.R
\name{summary.FilmRank}
\alias{summary.FilmRank}
\title{summary function for FilmRank results}
\usage{
\method{summary}{FilmRank}(filmrank.obj)
}
\arguments{
\item{filmrank.obj}{a FilmRank object}
}
\description{
\code{summary.FilmRank} prints the highest ranked film from each category.
}
