% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagff.R
\name{summary.LAGFF}
\alias{summary.LAGFF}
\title{summary function for LAGFF results}
\usage{
\method{summary}{LAGFF}(lagff.obj)
}
\arguments{
\item{lagff.obj}{an LAGFF object}
}
\description{
\code{summary.LAGFF} prints the highest ranked film from each category.
}
