% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagff.R
\name{plot.LAGFF}
\alias{plot.LAGFF}
\title{Plot LAGFF rankings}
\usage{
\method{plot}{LAGFF}(lagff.obj, rankmetric = "authority", color.palette,
  color.alpha = 1, title.colname, type.colname, title = NULL,
  cex.label = 0.5, xlim = c(1, 5), cex.xlab = 1, cex.ticks = 0.5,
  boxsize = 1/200, boxcolor = "darkblue", new_page = TRUE, grid = TRUE,
  xticks = seq(1, 5, by = 1))
}
\description{
\code{plot.LAGFF} creates a forest plot of LAGFF rankings.
}
